[#1] : Separated animation class into its own file

+++ src/Player.cpp
@@ -25,7 +25,7 @@ void PlayerObject::ProcessInput() {
 }
 
 void PlayerObject::Process() {
-	m_location += m_movementAccum * TIMER.GetFrameDuration();
+	m_location += m_movementAccum * (float)TIMER.GetFrameDuration();
 	if (!m_movementAccum.IsZero()) 
 		m_animation.Update();
